+++ .happy/terraform/envs/dev/main.tf
@@ -1,13 +1,16 @@
 module "stack" {
-  source           = "git@github.com:chanzuckerberg/happy//terraform/modules/happy-stack-eks?ref=main"
+  source           = "git@github.com:chanzuckerberg/happy//terraform/modules/happy-stack-eks?ref=happy-stack-eks-v4.27.1"
   image_tag        = var.image_tag
   image_tags       = jsondecode(var.image_tags)
   stack_name       = var.stack_name
   deployment_stage = "dev"
   stack_prefix     = "/${var.stack_name}"
   k8s_namespace    = "czid-dev-happy-happy-env"
+  app_name         = var.app
   additional_env_vars = {
     API_URL = "https://sandbox.czid.org"
+    NEXTGEN_ENTITIES_URL = "http://ryan-test-entities.czid-dev-happy-happy-env.svc.cluster.local:8008"
+    NEXTGEN_WORKFLOWS_URL = "http://workflows.czidnet"
   }
   services = {
     gql = {